Move remaining CES-related components to @woocommerce/customer-effort-score
