+++ packages/js/customer-effort-score/src/components/customer-effort-score-tracks-container/index.js
@@ -4,17 +4,15 @@
 import { useEffect } from 'react';
 import { compose } from '@wordpress/compose';
 import { withDispatch, withSelect } from '@wordpress/data';
-import {
-	QUEUE_OPTION_NAME,
-	STORE_KEY,
-} from '@woocommerce/customer-effort-score';
+import { createElement } from '@wordpress/element';
 import { OPTIONS_STORE_NAME } from '@woocommerce/data';
 import PropTypes from 'prop-types';
 
 /**
  * Internal dependencies
  */
-import CustomerEffortScoreTracks from './customer-effort-score-tracks';
+import { CustomerEffortScoreTracks } from '../';
+import { QUEUE_OPTION_NAME, STORE_KEY } from '../../store';
 
 /**
  * Maps the queue of CES tracks surveys to CustomerEffortScoreTracks
