+++ packages/js/customer-effort-score/src/constants.js
@@ -1,1 +1,7 @@
+export const ADMIN_INSTALL_TIMESTAMP_OPTION_NAME =
+	'woocommerce_admin_install_timestamp';
+
 export const ALLOW_TRACKING_OPTION_NAME = 'woocommerce_allow_tracking';
+
+export const SHOWN_FOR_ACTIONS_OPTION_NAME =
+	'woocommerce_ces_shown_for_actions';
