+++ packages/js/customer-effort-score/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-customer-effort-score-exit-page-tracker';
