+++ packages/js/customer-effort-score/src/index.ts
@@ -1,9 +1,5 @@
-export * from './components/customer-effort-score';
-export * from './components/customer-feedback-simple';
-export * from './components/customer-feedback-modal';
-export * from './components/product-mvp-feedback-modal';
-export * from './components/feedback-modal';
-export * from './hooks/use-customer-effort-score-exit-page-tracker';
-export * from './store';
-export * from './utils/customer-effort-score-exit-page';
+export * from './components';
 export * from './constants';
+export * from './hooks';
+export * from './store';
+export * from './utils';
