+++ packages/js/customer-effort-score/src/utils/get-store-age-in-weeks.ts
