+++ packages/js/customer-effort-score/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './customer-effort-score-exit-page';
+export * from './get-store-age-in-weeks';
