--- plugins/woocommerce-admin/client/customer-effort-score-tracks/constants.ts
@@ -1,4 +1,0 @@
-export const SHOWN_FOR_ACTIONS_OPTION_NAME =
-	'woocommerce_ces_shown_for_actions';
-export const ADMIN_INSTALL_TIMESTAMP_OPTION_NAME =
-	'woocommerce_admin_install_timestamp';
