--- plugins/woocommerce-admin/client/customer-effort-score-tracks/index.js
@@ -1,3 +1,0 @@
-export { default as CustomerEffortScoreTracks } from './customer-effort-score-tracks';
-export { default as CustomerEffortScoreTracksContainer } from './customer-effort-score-tracks-container';
-export * from './customer-effort-score-modal-container.tsx';
