+++ plugins/woocommerce-admin/client/homescreen/constants.js
@@ -9,9 +9,3 @@ export const WELCOME_MODAL_DISMISSED_OPTION_NAME =
  */
 export const WELCOME_FROM_CALYPSO_MODAL_DISMISSED_OPTION_NAME =
 	'woocommerce_welcome_from_calypso_modal_dismissed';
-
-/**
- * WooCommerce Admin installation timestamp option name.
- */
-export const WOOCOMMERCE_ADMIN_INSTALL_TIMESTAMP_OPTION_NAME =
-	'woocommerce_admin_install_timestamp';
